Delete ticket functionality
