+++ src/models/mod.rs
@@ -1,5 +1,5 @@
-mod ticket_draft;
 mod ticket;
+mod ticket_draft;
 
 pub use ticket::*;
 pub use ticket_draft::*;
